Changed how the weapon scripting is done. It now uses scriptable objects, so that it is easier to change values. Organised scripts into folders.
