+++ Assets/Items.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 632c5fee1c1d5ce4ca0fdf9ce5c226c5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
