+++ Assets/Items/Assault Rifle.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f07ed8fadc5211d47885c7f535b3d0f8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
