+++ Assets/Items/Claws.asset
@@ -1,0 +1,22 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ef411e071e5a68f47af6ef68ffc6a828, type: 3}
+  m_Name: Claws
+  m_EditorClassIdentifier: 
+  fireMode: 0
+  magCount: 0
+  fireRate: 0
+  magSize: 0
+  range: 0
+  damage: 0
+  magsLeft: 0
+  bulletsInCurrentMag: 0
