+++ Assets/Items/Claws.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c97386eb027242b41992e743522b1ddf
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
