+++ Assets/Items/SMG.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b634d8000c9a5094085804eb3b7a5c36
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
