+++ Assets/Items/Shotgun.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b689f39db0eda704bb22bed598c0e418
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
