+++ Assets/Resources/Alien (Cylinder).prefab
@@ -477,6 +477,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthSlider: {fileID: 6329036609802932500}
   maxHealth: 100
+  currentWeapon: {fileID: 11400000, guid: c97386eb027242b41992e743522b1ddf, type: 2}
 --- !u!1 &4418398108616668018
 GameObject:
   m_ObjectHideFlags: 0
