+++ Assets/Scripts/Alien/AlienController.cs
@@ -24,7 +24,7 @@ public class AlienController : AlienMovement
         alienInteraction = new PlayerInteraction();
         alienAttack = GetComponent<PlayerAttack>();
 
-        alienAttack.weaponDict.TryGetValue("claws", out alienAttack.currentWeapon);
+        //alienAttack.weaponDict.TryGetValue("claws", out alienAttack.currentWeapon);
     }
 
     private new void Update()
