+++ Assets/Scripts/General Scripts/Weapon.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a536d577bf921a44ba5399b5c0abafa1
+guid: ef411e071e5a68f47af6ef68ffc6a828
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
