+++ Assets/Scripts/Marine/FlashlightControl.cs
