+++ Assets/Scripts/Marine/MuzzleFlashScript.cs
