+++ Assets/Scripts/Miscellaneous/DoorTriggerScript.cs
