+++ Assets/Scripts/Miscellaneous/DoorTriggerScript.cs.meta
